% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_films.R
\name{rank_films}
\alias{rank_films}
\title{Rank Species by specified metric}
\usage{
rank_films(interested = NULL)
}
\arguments{
\item{interested}{Metric to rank films on. ("characters", "planets", "starships", "vehicles", "species")}
}
\value{
A plot with film name on the y-axis and the ranking metric on the x-axis.
}
\description{
Returns a plot showing all six Star Wars films
sorted by the count of a specified metric with the film name on the y-axis,
and the ranking metric on the x-axis.
}
\examples{
\dontrun{rank_films(interested = "species"),
rank_films(interested = "characters")}
}
