% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_starships-Ex.R, R/rank_starships.R
\name{rank_starships}
\alias{rank_starships}
\title{Rank Starships by specified metric}
\usage{
rank_starships(format = NULL, sort_on = NULL)

rank_starships(format = NULL, sort_on = NULL)
}
\value{



}
\description{
Rank Starships by specified metric

Rank Starships by specified metric
}
