% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_people.R
\name{rank_people}
\alias{rank_people}
\title{Rank Peoples by specified metric}
\usage{
rank_people(interested = NULL, n = 15)
}
\arguments{
\item{interested}{Metric to rank people on. ("height", "mass")}

\item{n}{Number of results to plot. (default = 15)}
}
\value{
A plot with people name on the y-axis and the ranking metric on the x-axis.
}
\description{
Returns a plot showing the top "n" number of Star Wars people
sorted by a specified metric with the person's name on the y-axis,
and the ranking metric on the x-axis.
}
\examples{
\dontrun{rank_people(interested = "height"),
rank_people(interested = "mass")}
}
