% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_planets.R
\name{rank_planets}
\alias{rank_planets}
\title{Rank Planets by quantitative metric (rotation_period, orbital_period, diameter, population)}
\usage{
rank_planets(interested = NULL, n = 15)
}
\arguments{
\item{interested}{Metric to rank planets on. ("rotation_period", "orbital_period", "diameter", "population")}

\item{n}{Number of results to plot. (default = 15)}
}
\value{
A ggplot with planet name on  the y-axis and the ranking metric on the x-axis.
}
\description{
Returns a plot showing the top "n" number of Star Wars planets
sorted by a specified metric with planet name on the y-axis,
and the ranking metric on the x-axis.
}
\examples{
\dontrun{rank_planets(interested = "rotation_period"),
rank_planets(interested = "population")}
}
