% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_planets.R
\name{rank_planets}
\alias{rank_planets}
\title{Rank Planets by quantitative metric (rotation_period, orbital_period, diameter, population)}
\usage{
rank_planets(format = NULL)
}
\value{

}
\description{
Rank Planets by quantitative metric (rotation_period, orbital_period, diameter, population)
}
