% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_starships.R
\name{rank_starships}
\alias{rank_starships}
\title{Rank Starships by specified metric}
\usage{
rank_starships(interested = NULL)
}
\arguments{
\item{interested}{Metric to rank starships on. ("cost_in_credits", "length", "max_atmosphering_speed", "crew", "passengers", "cargo_capacity")}
}
\value{
A plot with starship name on the y-axis and the ranking metric on  the x-axis.
}
\description{
Returns a plot showing the top 15 Star Wars starships
sorted by a specified metric with starship name on the y-axis,
and the ranking metric on the x-axis.
}
\examples{
\dontrun{rank_starships(interested = "crew"),
rank_starships(interested = "max_atmosphering_speed")}
}
